refactor: refactoring the name of the class Adress to Address
